# src/app/api/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "DealVault API",
    "version": "1.0.0",
    "description": "Commodity deal room management platform API. All endpoints require authentication via NextAuth session unless noted otherwise."
  },
  "servers": [
    { "url": "http://localhost:3000", "description": "Development" }
  ],
  "components": {
    "securitySchemes": {
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "next-auth.session-token"
      }
    },
    "schemas": {
      "Deal": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "dealNumber": { "type": "string", "example": "DV-2026-0001" },
          "title": { "type": "string" },
          "commodity": { "type": "string", "enum": ["gold", "diamonds", "platinum"] },
          "quantity": { "type": "number" },
          "unit": { "type": "string" },
          "value": { "type": "number" },
          "currency": { "type": "string", "default": "USD" },
          "commissionPool": { "type": "number", "minimum": 0, "maximum": 1 },
          "status": { "$ref": "#/components/schemas/DealStatus" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "DealStatus": {
        "type": "string",
        "enum": ["draft", "documents_pending", "under_review", "verified", "in_progress", "settled", "closed", "cancelled"]
      },
...
      "Notification": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "type": { "type": "string" },
          "title": { "type": "string" },
          "message": { "type": "string" },
          "read": { "type": "boolean" },
          "dealId": { "type": "string", "format": "uuid", "nullable": true },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
...
      "Pagination": {
        "type": "object",
        "properties": {
          "page": { "type": "integer" },
          "limit": { "type": "integer" },
          "total": { "type": "integer" },
          "totalPages": { "type": "integer" }
        }
      },
...
    }
  },
  "security": [{ "session": [] }],
  "paths": {
...
    "/api/deals": {
      "get": {
        "tags": ["Deals"],
        "summary": "List deals (with search, filters, pagination)",
        "parameters": [
          { "name": "search", "in": "query", "schema": { "type": "string" }, "description": "Search by title or deal number" },
          { "name": "status", "in": "query", "schema": { "$ref": "#/components/schemas/DealStatus" } },
          { "name": "commodity", "in": "query", "schema": { "type": "string", "enum": ["gold", "diamonds", "platinum"] } },
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 20, "maximum": 50 } }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of deals",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "deals": { "type": "array", "items": { "$ref": "#/components/schemas/Deal" } },
                    "pagination": { "$ref": "#/components/schemas/Pagination" }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Deals"],
        "summary": "Create a new deal room",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["title", "commodity", "quantity", "unit", "value"],
                "properties": {
                  "title": { "type": "string", "minLength": 3 },
                  "commodity": { "type": "string", "enum": ["gold", "diamonds", "platinum"] },
                  "quantity": { "type": "number", "exclusiveMinimum": 0 },
                  "unit": { "type": "string", "minLength": 1 },
                  "value": { "type": "number", "exclusiveMinimum": 0 },
                  "currency": { "type": "string", "default": "USD" },
                  "commissionPool": { "type": "number", "minimum": 0, "maximum": 1, "default": 0.02 }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Deal created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Deal" } } } },
          "400": { "description": "Validation error" }
        }
      }
    },
...
    "/api/deals/{id}/settle": {
      "post": {
        "tags": ["Deals"],
        "summary": "Initiate settlement (all parties must be accepted)",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "Deal settled, commission ledger entries created" },
          "400": { "description": "Not all parties accepted or invalid state" }
        }
      }
    },
...
    "/api/documents": {
      "post": {
        "tags": ["Documents"],
        "summary": "Upload a document to a deal",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": ["file", "dealId", "type"],
                "properties": {
                  "file": { "type": "string", "format": "binary" },
                  "dealId": { "type": "string" },
                  "type": { "type": "string", "enum": ["SPA", "NCNDA", "IMFPA", "BCL", "POF", "FCO", "ICPO", "LOI", "KYC", "AML", "other"] },
                  "visibility": { "type": "string", "enum": ["deal", "side", "private"], "default": "deal" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Document uploaded" },
          "400": { "description": "File validation failed (type/size)" }
        }
      }
    },
    "/api/documents/{id}": {
      "get": {
        "tags": ["Documents"],
        "summary": "Download a document (access control + visibility enforced)",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "File download stream" },
          "403": { "description": "Access denied" },
          "404": { "description": "Document not found" }
        }
      }
    },
...
    "/api/notifications": {
      "get": {
        "tags": ["Notifications"],
        "summary": "List notifications with unread count",
        "responses": {
          "200": {
            "description": "Notifications and unread count",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "notifications": { "type": "array", "items": { "$ref": "#/components/schemas/Notification" } },
                    "unreadCount": { "type": "integer" }
                  }
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": ["Notifications"],
        "summary": "Mark notifications as read",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "id": { "type": "string", "description": "Single notification ID" },
                  "all": { "type": "boolean", "description": "Mark all as read" }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "Notifications marked as read" } }
      }
    }
  }
}